refac: make internal/app dir #5
